tests/wpt/web-platform-tests/css/css-anchor-position/position-fallback-basics.html

<!DOCTYPE html>
<title>Tests basics of the 'position-fallback' property</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-position-fallback">
<link rel="author" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<div id="container">
  <div id="target"></div>
</div>

<script>
// position-fallback: none | <dashed-ident>
test_valid_value('position-fallback', 'none');
test_valid_value('position-fallback', '--foo');
test_invalid_value('position-fallback', 'foo-bar');
test_invalid_value('position-fallback', '--foo --bar')
test_invalid_value('position-fallback', '--foo, --bar')
test_invalid_value('position-fallback', '100px');
test_invalid_value('position-fallback', '100%');

// Computed value: as specified
test_computed_value('position-fallback', 'none');
test_computed_value('position-fallback', '--foo');

// Initial: none
// Inherited: no
assert_not_inherited('position-fallback', 'none', '--foo');

// Animation type: discrete
test_no_interpolation({
  property: 'position-fallback',
  from: '--foo',
  to: 'none',
});
</script>
